Move process CSV code into input/process.rs
