+++ src/input.rs
@@ -11,6 +11,8 @@ pub mod agent;
 pub use agent::read_agents;
 pub mod commodity;
 pub use commodity::read_commodities;
+pub mod process;
+pub use process::read_processes;
 pub mod region;
 pub use region::read_regions;
